Accept 'latest' as valid Firefox versions
Make a curl call and scrape the HTTP directory listing
to find the latest version of firefox corresponding to:
latest, latest-esr, latest-beta.

It seems that 'latest' and friends are just symbolic
links to the current respective versions.

There *may* be an API endpoint to discover them without
scraping HTML, but I have not found it.